refactor: move fetchFirstWalletAddress to sagas/wallet
